\name{mile}
\alias{mile}
\docType{data}
\title{World records for the mile run}
\description{
  World record times for the mile run, 1861--2003.
}
\format{
  A data frame with 46 observations:
  \describe{
    \item{Year}{Year in which the record was set}
    \item{Time}{Running time, in seconds}
    \item{Name}{Name of person setting the record}
    \item{Country}{Country of residence of the record setter}
    \item{Place}{Place the record was set}
    \item{Gender}{Gender of the record holder}
  }
}
\source{
<<<<<<< HEAD
The original source is no longer available online.
=======
Data source: \url{
http://www.saunalahti.fi/~sut/eng/
}
>>>>>>> 4d1499ef27c6aa7a1d198fa46d47d7e3ab700783
}
\references{Weisberg, S. (2014). \emph{Applied Linear Regression}, 4th
edition. Hoboken NJ: Wiley.}
\examples{
head(mile)
}
\keyword{datasets}
